Paywalls components/file split (#4200)
Removes code we won't be using for some time and splits up some large
files into component files.
