+++ RevenueCatUI/Views/TemplatePackageSetting.swift
@@ -38,7 +38,9 @@ extension PaywallTemplate {
         case .template4: return .multiple
         case .template5: return .multiple
         case .template7: return .multiTier
+        #if PAYWALL_COMPONENTS
         case .templateComponents: return .multiple
+        #endif
         }
     }
