+++ Sources/Paywalls/Components/PaywallComponentBases.swift
@@ -4,9 +4,12 @@
 //
 //  Created by Josh Holtz on 6/11/24.
 //
+// swiftlint:disable all
 
 import Foundation
 import SwiftUI // TODO: This feels wrong
+
+#if PAYWALL_COMPONENTS
 
 public typealias TierId = String
 public typealias LocaleId = String
@@ -31,11 +34,6 @@ public enum PaywallComponent: Decodable, Sendable, Hashable, Equatable {
     case tierToggle(TierToggleComponent)
     case text(TextComponent)
     case image(ImageComponent)
-    case video(VideoComponent)
-    case carousel(CarouselComponent)
-    case packages(PackagesComponent)
-    case features(FeaturesComponent)
-    case purchaseButton(PurchaseButtonComponent)
     case spacer(SpacerComponent)
 
     enum CodingKeys: String, CodingKey {
@@ -48,11 +46,6 @@ public enum PaywallComponent: Decodable, Sendable, Hashable, Equatable {
         case tierToggle
         case text
         case image
-        case video
-        case carousel
-        case packages
-        case features
-        case purchaseButton
         case spacer
     }
 
@@ -71,16 +64,6 @@ public enum PaywallComponent: Decodable, Sendable, Hashable, Equatable {
             self = .text(try TextComponent(from: decoder))
         case .image:
             self = .image(try ImageComponent(from: decoder))
-        case .video:
-            self = .video(try VideoComponent(from: decoder))
-        case .carousel:
-            self = .carousel(try CarouselComponent(from: decoder))
-        case .packages:
-            self = .packages(try PackagesComponent(from: decoder))
-        case .features:
-            self = .features(try FeaturesComponent(from: decoder))
-        case .purchaseButton:
-            self = .purchaseButton(try PurchaseButtonComponent(from: decoder))
         case .spacer:
             self = .spacer(try SpacerComponent(from: decoder))
         }
@@ -98,16 +81,6 @@ public enum PaywallComponent: Decodable, Sendable, Hashable, Equatable {
             return component.displayPreferences
         case .image(let component):
             return component.displayPreferences
-        case .video(let component):
-            return component.displayPreferences
-        case .carousel(let component):
-            return component.displayPreferences
-        case .packages(let component):
-            return component.displayPreferences
-        case .features(let component):
-            return component.displayPreferences
-        case .purchaseButton(let component):
-            return component.displayPreferences
         case .spacer(let component):
             return component.displayPreferences
         }
@@ -124,16 +97,6 @@ public enum PaywallComponent: Decodable, Sendable, Hashable, Equatable {
         case .text(let component):
             return component.focusIdentifiers
         case .image(let component):
-            return component.focusIdentifiers
-        case .video(let component):
-            return component.focusIdentifiers
-        case .carousel(let component):
-            return component.focusIdentifiers
-        case .packages(let component):
-            return component.focusIdentifiers
-        case .features(let component):
-            return component.focusIdentifiers
-        case .purchaseButton(let component):
             return component.focusIdentifiers
         case .spacer(let component):
             return component.focusIdentifiers
@@ -307,3 +270,5 @@ public extension PaywallComponent {
         }
     }
 }
+
+#endif
