+++ Sources/Paywalls/Components/PaywallPurchaseButtonComponent.swift
@@ -6,6 +6,9 @@
 //
 
 import Foundation
+// swiftlint:disable all
+
+#if PAYWALL_COMPONENTS
 
 public extension PaywallComponent {
     struct PurchaseButtonComponent: PaywallComponentBase {
@@ -33,3 +36,5 @@ public extension PaywallComponent {
 
     }
 }
+
+#endif
