+++ Sources/Paywalls/Components/PaywallSpacerComponent.swift
@@ -6,6 +6,9 @@
 //
 
 import Foundation
+// swiftlint:disable all
+
+#if PAYWALL_COMPONENTS
 
 public extension PaywallComponent {
     struct SpacerComponent: PaywallComponentBase {
@@ -24,3 +27,5 @@ public extension PaywallComponent {
 
     }
 }
+
+#endif
