+++ Sources/Paywalls/Components/PaywallTiersComponent.swift
@@ -6,6 +6,9 @@
 //
 
 import Foundation
+// swiftlint:disable all
+
+#if PAYWALL_COMPONENTS
 
 public extension PaywallComponent {
     struct TiersComponent: PaywallComponentBase {
@@ -97,3 +100,5 @@ public extension PaywallComponent {
 
     }
 }
+
+#endif
