--- Sources/Paywalls/Components/PaywallToggleComponent.swift
@@ -1,8 +1,0 @@
-//
-//  File.swift
-//  
-//
-//  Created by Josh Holtz on 6/14/24.
-//
-
-import Foundation
